refactor(core-interfaces): organise core-p2p interfaces
