+++ packages/core-interfaces/src/core-p2p/peer-guard.ts
@@ -1,0 +1,37 @@
+import { Dato } from "@faustbrian/dato";
+import { IPeer } from "./peer";
+
+export interface IPeerGuard {
+    punishment(offence: string): IPunishment;
+    analyze(peer: IPeer): IPunishment;
+    isWhitelisted(peer: IPeer): boolean;
+    isValidVersion(peer: IPeer): boolean;
+    isValidNetwork(peer: IPeer): boolean;
+    isValidPort(peer: IPeer): boolean;
+}
+
+export interface IPeerSuspension {
+    readonly peer: any;
+    readonly punishment: IPunishment;
+
+    nextSuspensionReminder?: Dato;
+
+    isLow(): boolean;
+    isMedium(): boolean;
+    isHigh(): boolean;
+    isCritical(): boolean;
+
+    hasExpired(): boolean;
+}
+
+export interface IOffence {
+    until: () => Dato;
+    reason: string;
+    severity?: "low" | "medium" | "high" | "critical";
+}
+
+export interface IPunishment {
+    until: Dato;
+    reason: string;
+    severity?: "low" | "medium" | "high" | "critical";
+}
