+++ packages/core-interfaces/src/core-p2p/peer-processor.ts
@@ -1,0 +1,16 @@
+import { IPeer } from "./peer";
+import { IPunishment } from "./peer-guard";
+
+export interface IAcceptNewPeerOptions {
+    seed?: boolean;
+    lessVerbose?: boolean;
+}
+
+export interface IPeerProcessor {
+    validateAndAcceptPeer(peer, options?: IAcceptNewPeerOptions): Promise<void>;
+    validatePeer(peer, options?: IAcceptNewPeerOptions): boolean;
+    acceptNewPeer(peer, options?: IAcceptNewPeerOptions): Promise<void>;
+    suspend(peer: IPeer, punishment?: IPunishment): void;
+    unsuspend(peer: IPeer): Promise<void>;
+    isSuspended(peer: IPeer): boolean;
+}
