+++ packages/core-interfaces/src/core-p2p/peer-service.ts
@@ -1,0 +1,15 @@
+import { INetworkMonitor } from "./network-monitor";
+import { IPeerCommunicator } from "./peer-communicator";
+import { IPeerConnector } from "./peer-connector";
+import { IPeerGuard } from "./peer-guard";
+import { IPeerProcessor } from "./peer-processor";
+import { IPeerStorage } from "./peer-storage";
+
+export interface IPeerService {
+    getStorage(): IPeerStorage;
+    getProcessor(): IPeerProcessor;
+    getConnector(): IPeerConnector;
+    getCommunicator(): IPeerCommunicator;
+    getMonitor(): INetworkMonitor;
+    getGuard(): IPeerGuard;
+}
