+++ packages/core-interfaces/src/core-p2p/peer-verifier.ts
@@ -1,0 +1,6 @@
+export interface IPeerVerificationResult {
+    readonly myHeight: number;
+    readonly hisHeight: number;
+    readonly highestCommonHeight: number;
+    readonly forked: boolean;
+}
