+++ packages/core-interfaces/src/core-p2p/peer.ts
@@ -1,0 +1,43 @@
+import { Dato } from "@faustbrian/dato";
+import { IPeerVerificationResult } from "./peer-verifier";
+
+export interface IPeer {
+    readonly url: string;
+
+    ip: string;
+    port: number;
+
+    nethash: string;
+    version: string;
+    os: string;
+
+    delay: number;
+    downloadSize: number;
+    headers: Record<string, string | number>;
+    state: any; // @TODO: add an interface/type
+    lastPinged: Dato | null;
+    verificationResult: IPeerVerificationResult | null;
+
+    // @TODO: review and remove them where appropriate
+    status: any;
+    commonBlocks: any;
+    socketError: any;
+
+    setHeaders(headers: Record<string, string>): void;
+
+    isVerified(): boolean;
+    isForked(): boolean;
+    recentlyPinged(): boolean;
+
+    toBroadcast(): IPeerBroadcast;
+}
+
+export interface IPeerBroadcast {
+    ip: string;
+    port: number;
+    nethash: string;
+    version: string;
+    os: string;
+    height: number;
+    delay: number;
+}
