+++ packages/core-interfaces/src/core-p2p/server.ts
@@ -1,0 +1,23 @@
+import { models } from "@arkecosystem/crypto";
+import { IDelegateWallet } from "../core-database";
+
+export interface IResponse<T> {
+    data: T;
+}
+
+export interface ICurrentRound {
+    current: number;
+    reward: string;
+    timestamp: number;
+    delegates: IDelegateWallet[];
+    currentForger: IDelegateWallet;
+    nextForger: IDelegateWallet;
+    lastBlock: models.IBlockData;
+    canForge: boolean;
+}
+
+export interface IForgingTransactions {
+    transactions: string[];
+    poolSize: number;
+    count: number;
+}
